Cap edu/backend/requirements.txt
# Requirements for QazKids Backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
bcrypt==4.1.2
pyjwt==2.8.1
requests==2.31.0
